Refactor find_key and find_key_path helpers
